pi5_server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
RPi.GPIO==0.7.1
firebase-admin==6.3.0
picamera2==0.3.16
opencv-python==4.8.1.78
numpy==1.24.3
